requirements.txt
aiodns==1.1.1
aiohttp==3.6.2
alabaster==0.7.12
anaconda-client==1.7.2
anaconda-project==0.8.3
APScheduler==3.6.3
argh==0.26.2
asn1crypto==1.3.0
astroid
astropy==4.0
async-timeout==3.0.1
atomicwrites==1.3.0
attrs==19.3.0
autobahn==20.3.1
Automat==20.2.0
autopep8==1.4.4
Babel==2.8.0
backcall==0.1.0
backports.os==0.1.1
backports.shutil-get-terminal-size==1.0.0
bcrypt==3.1.7
beautifulsoup4==4.8.2
bitarray==1.2.1
bkcharts==0.2
bleach==3.1.0
bokeh==2.0.0
boto==2.49.0
Bottleneck==1.3.2
ccxt==1.24.74
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
click==7.1.1
cloudpickle==1.3.0
clyent==1.2.2
colorama==0.4.3
colorlog==4.1.0
comtypes==1.1.7
constantly==15.1.0
contextlib2==0.6.0.post1
cryptography==2.8
cssselect==1.1.0
cycler==0.10.0
Cython==0.29.15
cytoolz==0.10.1
dask==2.12.0
dateparser==0.7.4
decorator==4.4.2
defusedxml==0.6.0
diff-match-patch==20181111
distributed==2.12.0
dnspython==1.16.0
docutils==0.16
entrypoints==0.3
et-xmlfile==1.0.1
eventlet==0.25.1
fastcache==1.1.0
feedfinder2==0.0.4
feedparser==5.2.1
filelock==3.0.12
flake8==3.7.9
Flask==1.1.1
Flask-Cache==0.13.1
Flask-Caching==1.8.0
flask-mongoengine==0.9.5
Flask-PyMongo==2.3.0
Flask-SocketIO==4.2.1
Flask-WTF==0.14.3
fsspec==0.6.2
future==0.18.2
gevent==1.4.0
glob2==0.7
greenlet==0.4.15
h5py==2.10.0
HeapDict==1.0.1
html5lib==1.0.1
hyperlink==19.0.0
hypothesis==5.5.4
idna==2.9
imageio==2.8.0
imagesize==1.2.0
importlib-metadata==1.5.0
incremental==17.5.0
intervaltree==3.0.2
ipykernel==5.1.4
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.15.2
jieba3k==0.35.1
Jinja2==2.11.1
joblib==0.14.1
json5==0.9.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.0.0
jupyter-console==6.1.0
jupyter-core==4.6.1
jupyterlab==1.2.6
jupyterlab-server==1.0.7
keyring==21.1.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
libarchive-c==2.8
llvmlite==0.31.0
locket==0.2.0
lxml==4.5.0
MarkupSafe==1.1.1
matplotlib==3.1.3
mccabe==0.6.1
menuinst==1.4.16
mistune==0.8.4
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
mock==4.0.1
mongoengine==0.19.1
monotonic==1.5
more-itertools==8.2.0
mpmath==1.1.0
msgpack==0.6.1
multidict==4.7.5
multipledispatch==0.6.0
nbconvert==5.6.1
nbformat==5.0.4
networkx==2.4
newspaper3k==0.2.8
nltk==3.4.5
nose==1.3.7
notebook==6.0.3
numba==0.48.0
numexpr==2.7.1
numpy==1.18.1
numpydoc==0.9.2
oauthlib==3.1.0
OctoBot-Tentacles-Manager==1.0.13
olefile==0.46
openpyxl==3.0.3
packaging==20.3
pandas==1.0.2
pandocfilters==1.4.2
paramiko==2.7.1
parso==0.5.2
partd==1.1.0
path==13.1.0
pathlib2==2.3.5
pathtools==0.1.2
patsy==0.5.1
pep8==1.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
pkginfo==1.5.0.1
pluggy==0.13.1
ply==3.11
praw==6.5.1
prawcore==1.0.1
prometheus-client==0.7.1
prompt-toolkit==3.0.3
psutil==5.7.0
py==1.8.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycares==3.1.1
pycodestyle==2.5.0
pycosat==0.6.3
pycparser==2.20
pycrypto==2.6.1
pycurl==7.43.0.5
pydocstyle==4.0.1
pyflakes==2.1.1
Pygments==2.6.1
PyHamcrest==2.0.2
pylint==2.4.4
pymongo==3.10.1
PyNaCl==1.3.0
pyodbc===4.0.0-unsupported
pyOpenSSL==19.1.0
pyparsing==2.4.6
pyreadline==2.1
pyrsistent==0.15.7
PySocks==1.7.1
pytest==5.4.1
pytest-arraydiff==0.3
pytest-astropy==0.8.0
pytest-astropy-header==0.1.2
pytest-doctestplus==0.5.0
pytest-openfiles==0.4.0
pytest-remotedata==0.3.2
python-binance==0.7.3
python-dateutil==2.8.1
python-engineio==3.12.1
python-jsonrpc-server==0.3.4
python-language-server==0.31.9
python-socketio==4.5.1
python-telegram-bot==12.4.2
python-twitter==3.5
pytz==2019.3
PyWavelets==1.1.1
pywin32==227
pywin32-ctypes==0.2.0
pywinpty==0.5.7
PyYAML
pyzmq==18.1.1
QDarkStyle==2.8
QtAwesome==0.7.0
qtconsole==4.7.1
QtPy==1.9.0
regex==2020.2.20
requests==2.23.0
requests-file==1.4.3
requests-oauthlib==1.3.0
rope==0.16.0
Rtree==0.9.3
ruamel-yaml==0.15.87
scikit-image==0.16.2
scikit-learn==0.22.1
scipy==1.4.1
seaborn==0.10.0
Send2Trash==1.5.0
service-identity==18.1.0
simplegeneric==0.8.1
simplifiedpytrends==1.0.4
singledispatch==3.4.0.3
six==1.14.0
snowballstemmer==2.0.0
sortedcollections==1.1.2
sortedcontainers==2.1.0
soupsieve==2.0
Sphinx==2.4.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sphinxcontrib-websupport==1.2.0
spyder==4.1.1
spyder-kernels==1.9.0
SQLAlchemy==1.3.15
statsmodels==0.11.0
sympy==1.5.1
tables==3.6.1
tblib==1.6.0
terminado==0.8.3
testpath==0.4.4
tinysegmenter==0.3
tldextract==2.2.2
toolz==0.10.0
tornado==6.0.4
tqdm==4.43.0
traitlets==4.3.3
tulipy==0.4.0
Twisted==20.3.0
txaio==20.1.1
typing-extensions==3.7.4.1
tzlocal==2.0.0
ujson==1.35
unicodecsv==0.14.1
update-checker==0.16
urllib3==1.25.3
vaderSentiment==3.3.1
watchdog==0.10.2
wcwidth==0.1.8
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==1.0.0
widgetsnbextension==3.5.1
win-inet-pton==1.1.0
win-unicode-console==0.5
wincertstore==0.2
wrapt==1.12.1
WTForms==2.2.1
xlrd==1.2.0
XlsxWriter==1.2.8
xlwings==0.18.0
xlwt==1.3.0
yapf==0.28.0
yarl==1.1.0
zict==2.0.0
zipp==2.2.0
zope.interface==5.0.0
